chore: moved my access related blocs to folder directory
