+++ lib/app/feature/my_access/my_access_change_password/my_access_change_password_screen.dart
@@ -7,7 +7,7 @@ import 'package:rgs_hris/app/feature/dashboard/drawer_widget.dart';
 import 'package:rgs_hris/router/app_route.dart';
 
 import '../../../../core/ui/widget/text_form_field_widget.dart';
-import '../../../bloc/my_access_bloc.dart';
+import '../../../bloc/my_access/my_access_bloc.dart';
 import '../../../common/util/key_strings.dart';
 
 class MyAccessChangePasswordScreen extends StatelessWidget {
